+++ CHANGELOG.md
@@ -6,10 +6,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 
-## Added
+### Changed
 
+- `es-verifactu-v1`: `[]*org.Note` validation will check the note's length if `key=general`
+
+### Added
+
+- `it-sdi-v1`: Validate names and addresses contain only latin characters as required by SDI.
 - `it-sdi-v1`: Payment means now support `online` key as a card payment.
-- `it-sdi-v1`: Validate names and addresses contain only latin characters as required by SDI
 - `org`: `Unit` additional standard units.
 - `es/verifactu`: map identity keys for specific codes.
 - `org`: `Person` new `key`, `addresses`, and `identities` properties.